requirements.txt
fastapi
numpy==1.26.4
pillow==10.3.0
tensorflow==2.15.0